requirements.txt
aiofiles==24.1.0
aiogram>=3.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
magic-filter==1.0.12
multidict==6.3.2
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
requests>=2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3
pandas>=2.0.0
openpyxl>=3.1.0

